refactor functions to include init (#4)
This change introduces two large refactor efforts.

The first refactor introduces a new init function in inlcudes/ that is
used to validate the role has the necessary parameters to function.  All
functions have been updated to inlcude init.yaml

The second change involves the load_config function.  Mostly this is a
rearrange of the tasks.  It also breaks out the handling of
checkpointing the current configuration into a separate set of includes
for reuse later.